chore: rename module `next` -> `client`
Since it's no longer really the "next" API it's the current.   I like
the encapsulation of it though so I'm keeping that around for now.
